some reorganization of the code in guilerobots.scm
It was way much pain having all this code trying not
to add the imi libs a second time to the load path than
just adding it another time, even if it shouldn't be the
right path. So I just removed all this overcomplicated
code and replaced it with this dump version.